@@ -1,11 +1,10 @@
-﻿namespace System.RuntimeCompilerServices
-{
-    [AttributeUsage(AttributeTargets.Parameter, AllowMultiple = false, Inherited = false)]
-    internal sealed class CallerArgumentExpressionAttribute : Attribute
-    {
-        public string ParameterName { get; set; }
+﻿namespace System.RuntimeCompilerServices;
 
-        public CallerArgumentExpressionAttribute(string parameterName) =>
-            ParameterName = parameterName;
-    }
+[AttributeUsage(AttributeTargets.Parameter, AllowMultiple = false, Inherited = false)]
+internal sealed class CallerArgumentExpressionAttribute : Attribute
+{
+    public string ParameterName { get; set; }
+
+    public CallerArgumentExpressionAttribute(string parameterName) =>
+        ParameterName = parameterName;
 }
