@@ -2,12 +2,11 @@
 // Use of this source code is governed by the Apache License 2.0,
 // as found in the LICENSE.txt file.
 
-namespace CameraControl.Visca
+namespace CameraControl.Visca;
+
+public enum PowerStatus
 {
-    public enum PowerStatus
-    {
-        On = 2,
-        Standby = 3,
-        InternalPowerCircuitError = 4
-    }
+    On = 2,
+    Standby = 3,
+    InternalPowerCircuitError = 4
 }
