@@ -2,27 +2,24 @@
 // Use of this source code is governed by the Apache License 2.0,
 // as found in the LICENSE.txt file.
 
-using System;
+namespace CameraControl.Visca;
 
-namespace CameraControl.Visca
+/// <summary>
+/// Represents a transport-level failure in VISCA communications.
+/// This is not thrown when a camera reports an error.
+/// </summary>
+public class ViscaProtocolException : Exception
 {
-    /// <summary>
-    /// Represents a transport-level failure in VISCA communications.
-    /// This is not thrown when a camera reports an error.
-    /// </summary>
-    public class ViscaProtocolException : Exception
+    public ViscaProtocolException()
     {
-        public ViscaProtocolException()
-        {
-        }
+    }
 
-        public ViscaProtocolException(string message) : base(message)
-        {
-        }
+    public ViscaProtocolException(string message) : base(message)
+    {
+    }
 
-        public ViscaProtocolException(string message, Exception innerException)
-            : base(message, innerException)
-        {
-        }
+    public ViscaProtocolException(string message, Exception innerException)
+        : base(message, innerException)
+    {
     }
 }
