@@ -2,26 +2,23 @@
 // Use of this source code is governed by the Apache License 2.0,
 // as found in the LICENSE.txt file.
 
-using System;
+namespace CameraControl.Visca;
 
-namespace CameraControl.Visca
+/// <summary>
+/// Represents a failure successfully reported by a VISCA endpoint.
+/// </summary>
+public class ViscaResponseException : Exception
 {
-    /// <summary>
-    /// Represents a failure successfully reported by a VISCA endpoint.
-    /// </summary>
-    public class ViscaResponseException : Exception
+    public ViscaResponseException()
     {
-        public ViscaResponseException()
-        {
-        }
+    }
 
-        public ViscaResponseException(string message) : base(message)
-        {
-        }
+    public ViscaResponseException(string message) : base(message)
+    {
+    }
 
-        public ViscaResponseException(string message, Exception innerException)
-            : base(message, innerException)
-        {
-        }
+    public ViscaResponseException(string message, Exception innerException)
+        : base(message, innerException)
+    {
     }
 }
